refactor: moved user selectors to a separate module
