+++ src/helpers/RequireAuth.tsx
@@ -1,10 +1,10 @@
 import { ReactNode } from "react";
 import { Navigate } from "react-router";
-import { RootState, useAppSelector } from "../store/store";
 import { ROUTES } from "../constants/routes.constants";
+import { useSelectJwt } from "../store/user/user.selectors";
 
 export const RequireAuth = ({ children }: { children: ReactNode }) => {
-  const jwt = useAppSelector((state: RootState) => state.user.jwt);
+  const jwt = useSelectJwt();
 
   if (!jwt) {
     return <Navigate to={ROUTES.signin} replace />;
